Split ConvivaAnalytics
