+++ src/ts/ConvivaSsaiAnalytics.ts
@@ -1,0 +1,9 @@
+import { ConvivaAnalyticsTracker } from './ConvivaAnalyticsTracker';
+
+export class ConvivaSsaiAnalytics {
+  private readonly convivaAnalyticsTracker: ConvivaAnalyticsTracker;
+
+  constructor(convivaAnalyticsTracker: ConvivaAnalyticsTracker) {
+    this.convivaAnalyticsTracker = convivaAnalyticsTracker;
+  }
+}
